add entities and response DTOs
Add core entity models and DTO structures for intranet portal

+++ src/main/java/com/jforce/intranetportal/dto/request/EmployeeCreateRequest.java
@@ -33,5 +33,5 @@ public class EmployeeCreateRequest {
     private String password;
 
     @NotNull(message = "Department id cannot be null")
-    private Long departmentId;
+    private Integer departmentId;
 }
